revise the size type in SZ_endCmpr()

+++ zc/include/zc.h
@@ -275,7 +275,7 @@ long ZC_computeDataLength(size_t r5, size_t r4, size_t r3, size_t r2, size_t r1)
 
 ZC_DataProperty* ZC_startCmpr_offline(char* varName, int dataType, void* oriData, size_t r5, size_t r4, size_t r3, size_t r2, size_t r1);
 ZC_DataProperty* ZC_startCmpr_offline_withDataAnalysis(char* varName, int dataType, void *oriData, size_t r5, size_t r4, size_t r3, size_t r2, size_t r1);
-ZC_CompareData* ZC_endCmpr_offline(ZC_DataProperty* dataProperty, int cmprSize);
+ZC_CompareData* ZC_endCmpr_offline(ZC_DataProperty* dataProperty, long cmprSize);
 void ZC_startDec_offline();
 void ZC_endDec_offline(ZC_CompareData* compareResult, char* solution, void *decData);
